fix Hive tests

+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/grouping.scala
@@ -49,16 +49,18 @@ case class Rollup(groupByExprs: Seq[Expression]) extends GroupingSet {}
 case class Grouping(child: Expression) extends Expression with Unevaluable {
   override def references: AttributeSet = AttributeSet(VirtualColumn.groupingIdAttribute :: Nil)
   override def children: Seq[Expression] = child :: Nil
-  override def dataType: DataType = IntegerType
+  override def dataType: DataType = ByteType
   override def nullable: Boolean = false
 }
 
 /**
   * GroupingID is a function that computes the level of grouping.
+  *
+  * If groupByExprs is empty, it means all grouping expressions in GroupingSets.
   */
 case class GroupingID(groupByExprs: Seq[Expression]) extends Expression with Unevaluable {
   override def references: AttributeSet = AttributeSet(VirtualColumn.groupingIdAttribute :: Nil)
   override def children: Seq[Expression] = groupByExprs
-  override def dataType: DataType = ByteType
+  override def dataType: DataType = IntegerType
   override def nullable: Boolean = false
 }
